refactor: extract symbol index into SRP microcrate
