+++ crates/perl-lsp-tooling/Cargo.toml
@@ -25,6 +25,7 @@ perl-subprocess-runtime = { workspace = true }
 perl-tdd-support = { workspace = true }
 perl-parser-core = { workspace = true }
 moka = { version = "0.12", features = ["sync"] }
+perl-symbol-index = { workspace = true }
 lsp-types = { version = "0.97.0", optional = true }
 serde = { version = "1.0.228", features = ["derive"] }
